Fix auto-approve bar not working and centralize feature
